requirements/base.txt
# Django
Django==1.11.4

# Environmental var configuration
django-configurations==2.0
dj-database-url==0.4.2
django-cache-url==2.0.0

# For the persistence stores
psycopg2==2.7.1
django-redis==4.8.0
celery[eventlet, redis, sqs]==4.1.0

# Models
django-model-utils==3.0.0

# Rest apis
djangorestframework==3.6.4
#Markdown==2.6.8
#django-filter==1.0.4

# Static files
whitenoise==3.3.0

# Media files
#boto==2.47.0
#django_unique_upload==0.2.1
Pillow==4.2.1

# Image manipulation
#django-versatileimagefield==1.7.0

# For asynchronous queuing
#django-rq==0.9.5

# Time zones support
pytz==2017.2

# Parsing Secondlife api's
backoff==1.4.3
requests==2.18.4
beautifulsoup4==4.6.0
lxml==3.8.0
djproxy==2.3.4
